Articulos: ListarConImagenes , Categorias: correccion consulta listar()
+ agrego .gitignore

+++ TPC_Comercio_Equipo26/TPC_Equipo26/Dominio/Imagen.cs
@@ -7,10 +7,9 @@ namespace TPC_Equipo26.Dominio
 {
     public class Imagen
     {
-        public int ID {get; set;}
+        public long ID {get; set;}
         public int IdArticulo { get; set; }
         public string UrlImagen { get; set; }
-        
         public bool Activo { get; set; }
     }
 }
